raci-project-manager/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic>=0.34.0
openai==1.47.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
